Update DiskPlot to use real data from API
Removed an unnecessary space in Server.kt  and DiskPlot.kt was updated to fetch real data from the API. It now handles response data in the form of Disk class with attributes; space, spacePercentage, quota, and availableQuota. The computed space utilization and available quota are now displayed in the DonutChart.